+++ config/default.js
@@ -17,7 +17,6 @@ const
 	path = require('upath');
 
 const
-	camelize = require('camelize'),
 	o = require('@v4fire/config/options').option;
 
 const
@@ -171,7 +170,7 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 						return obj;
 					}
 
-					return [Object.isObject(obj) ? obj : {name: obj}];
+					return [Object.isDictionary(obj) ? obj : {name: obj}];
 
 				} catch {}
 
@@ -185,12 +184,12 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 						try {
 							const
 								demo = require(pzlr.resolve.blockSync(`${name}/demo.js`)),
-								suit = camelize(args['--suit'] || 'demo');
+								suit = (args['--suit'] || 'demo').camelize(false);
 
 							const
 								wrap = (d) => [].concat((d || []).map((p) => ({name, ...p})));
 
-							if (Object.isObject(demo)) {
+							if (Object.isDictionary(demo)) {
 								return wrap(demo[suit]);
 							}
 
@@ -533,7 +532,7 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 		 */
 		publicPath(...args) {
 			const
-				concatURLs = require('urlconcat').concat;
+				{concatURLs} = require('@v4fire/core/lib/core/url');
 
 			let pathVal = o('public-path', {
 				env: true,